pyproject.toml
[project]
name = "video-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "deep-translator>=1.11.4",
    "demucs>=4.0.1",
    "edge-srt-to-speech>=0.0.24",
    "gradio>=5.39.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
]
